Merged in jrussek/fixBrightnessCheck (pull request #1619)
Adjust behavior where FlatDevice brightness = 0

* Adjust behavior where FlatDevice brightness = 0

As per ASCOM v2 a brightness of 0 now indicates the flatdevice to be
turned off. As NINA adjust the Brightness to MinBrightness when it's
zero, this can have the unintentional side effect of turning a
flatdevice on when setting the brightness to 0.

With this patch, NINA only adjusts the brightness if the value is larger
than 0.

Additionally it fixes the SetBrightness sequencer action throwing an exception
when setting to a brightness outside of the MinBrightness <> MaxBrightness
range, as it checks if the resulting Brightness equals the original target.